refactor: 목데이터 폴더 분리
