+++ src/mocks/dashboard/user.mock.ts
@@ -1,12 +1,4 @@
-export type Permission = "전체 허용" | "편집 허용" | "읽기 허용";
-
-interface User {
-  id: number;
-  name: string;
-  email: string;
-  avatar: string;
-  permission: Permission;
-}
+import { User } from "@/types/dashboard/user.type";
 
 export const MOCK_DATA_USERS: User[] = [
   {
